feat: add reset user password

+++ app/modules/admin/settings/acl/users-management/usecases/delete-user/delete-user-usecase-impl.ts
@@ -1,6 +1,7 @@
 import { IEventDispatcher, Either, left, right } from 'app/core/domain'
 import { UserNotFoundError } from 'app/modules/@shared/domain/errors'
 import { DeleteUserUseCase, DeleteUserUseCaseInput } from '../../domain'
+import { RootUserCannotBeModified } from '../../domain/errors'
 import { UserDeletedEvent } from '../../domain/events/user-deleted-event'
 import { FindUsernameRepository, UpdateUserRepository } from '../block-user'
 
@@ -11,13 +12,17 @@ export class DeleteUserUseCaseImpl implements DeleteUserUseCase {
     private readonly eventDispatcher: IEventDispatcher
   ) {}
 
-  public async perform (input: DeleteUserUseCaseInput): Promise<Either<UserNotFoundError, boolean>> {
+  public async perform (input: DeleteUserUseCaseInput): Promise<
+  Either<UserNotFoundError | RootUserCannotBeModified, boolean>> {
     const userEntity = await this.findUserNameRepository.findUsername(input.username)
 
     if (!userEntity) {
       return left(new UserNotFoundError())
     }
 
+    if (userEntity.isRoot) {
+      return left(new RootUserCannotBeModified())
+    }
     userEntity.delete()
 
     await this.updateUserRepository.update(userEntity)
